No message specified

+++ cogs/moderation.py
@@ -1,4 +1,4 @@
-import discord, random, asyncio
+import discord, random, asyncio, aiofiles, json
 from discord.ext import commands
 
 # PurgeError
@@ -10,54 +10,60 @@ def is_bot(m):
 	return 	m.author.bot
 def is_not_bot(m):
 	return 	not(m.author.bot)
-async def purge_messages(number, channel, mode, check=None):
-	if check is None:
-		return await channel.purge(limit=number)
-	diff_message = 0
-	total_message = 0
-	async for message in channel.history(limit=None):
-		if diff_message == number:
-			break
-		if check(message):
-			diff_message += 1
-		total_message += 1
-	else:
-		e = PurgeError(f'Could not find enough messages with mode {mode}')
-		raise e
 
-	return await channel.purge(limit=total_message, check=check)
+async def readDB():
+	try:
+		async with aiofiles.open('/home/tyman/code/utilibot/data.json', mode='r') as f:
+			return json.loads(await f.read())
+	except Exception as e:
+		print(f"An error occured, {e}")
 
+async def writeDB(data: dict):
+	try:
+		async with aiofiles.open('/home/tyman/code/utilibot/data.json', mode='r') as f_main:
+			async with aiofiles.open('/home/tyman/code/utilibot/data.json.bak', mode='w') as f_bak:
+				await f_bak.write(await f_main.read())
+		async with aiofiles.open('/home/tyman/code/utilibot/data.json', mode='w') as f:
+			d = json.dumps(data)
+			await f.write(d)
+	except Exception as e:
+		print(f"An error occured, {e}")
 
 class Moderation(commands.Cog):
 	def __init__(self, bot):
 		self.bot = bot
 
-	@commands.command()
+	@commands.group(invoke_without_command=True)
 	@commands.has_permissions(manage_messages=True)
 	@commands.bot_has_permissions(manage_messages=True)
 	@commands.guild_only()
-	async def purge(self, ctx, number, mode="all"):
+	async def purge(self, ctx, number: int=10):
 		"""
 		Purge a specified amount of messages from the current channel.
 
-		Number = The number of messages to delete, depending on the mode. If the mode is all, it will just delete this number of messages. If the mode is bot, it will delete this number of messages made by bots.
-		Mode = The mode of deleteing messages, can be all (defualt), bot, or humans (opposite of bot)
+		Number = The number of messages to delete.
 		"""
-		await ctx.message.delete()
-		mode = str(mode).lower()
-		number = int(number)
-		try:
-			if mode == "all":
-				deleted = await purge_messages(number, ctx.channel, mode)
-			elif mode == "bot":
-				deleted = await purge_messages(number, ctx.channel, mode, is_bot)
-			elif mode == "human":
-				deleted = await purge_messages(number, ctx.channel, mode, is_not_bot)
-			else:
-				return await ctx.send('Mode must be one of: all (default), bot, or human')
-		except PurgeError as e:
-			return await ctx.send(e)
+		if ctx.invoked_subcommand is None:
+			async with ctx.typing():
+				await ctx.message.delete()
+				deleted = await ctx.channel.purge(limit=number)
+			message = await ctx.channel.send(f'Deleted {len(deleted)} message(s)')
+			await asyncio.sleep(2.5)
+			await message.delete()
 
+	@purge.command()
+	@commands.has_permissions(manage_messages=True)
+	@commands.bot_has_permissions(manage_messages=True)
+	@commands.guild_only()
+	async def bot(self, ctx, number:int =10):
+		"""
+		Purge a specified amount of messages from the current channel. It will only delete messages made by bots.
+
+		Number = The number of messages to delete.
+		"""
+		async with ctx.typing():
+			await ctx.message.delete()
+			deleted = await ctx.channel.purge(limit=number, check=is_bot)
 		message = await ctx.channel.send(f'Deleted {len(deleted)} message(s)')
 		await asyncio.sleep(2.5)
 		await message.delete()
@@ -83,6 +89,74 @@ class Moderation(commands.Cog):
 				await ctx.send(f"Eror: Could Not DM user")
 				await member.kick(reason=f"{member.name} was kicked by {ctx.author.name}, for the reason: {reason}")
 				await ctx.send(f"Kicked {member} for the reason: `{reason}`")
+
+	@commands.command(name="hardlock", aliases=['lockdown', 'hl', 'ld'])
+	@commands.bot_has_permissions(manage_channels=True)
+	@commands.has_permissions(manage_channels=True)
+	async def hardlock(self, ctx):
+		"""
+		Locks down a channel by denying @everyone send messages permission.
+		"""
+		perms = ctx.channel.overwrites_for(ctx.guild.default_role)
+		perms.send_messages = False
+		await ctx.channel.set_permissions(ctx.guild.default_role, overwrite=perms, reason="Muted!")
+		await ctx.send("Successfully locked down this channel by removing send messages permission for @everyone.", allowed_mentions=discord.AllowedMentions(everyone=False, users=False, roles=False))
+
+	@commands.command(name="softlock", aliases=['lock', 'sl'])
+	@commands.bot_has_permissions(manage_messages=True)
+	@commands.has_permissions(manage_messages=True)
+	async def softlock(self, ctx, channel: discord.TextChannel=None, *, reason=None):
+		"""
+		Locks down a channel by deleting messages that people send.
+		"""
+		db = await readDB()
+		ch = channel or ctx.channel
+		if not str(ch.id) in db["softlocked_channels"]:
+			db["softlocked_channels"][str(ch.id)] = {
+				"user": str(ctx.author.id),
+				"reason": reason or "No reason provided.",
+				"whitelist": []
+			}
+			await writeDB(db)
+			await ctx.send("Successfully softlocked.")
+		else:
+			await ctx.send("Channel is already softlocked.")
+
+	@commands.command(aliases=['wh'])
+	@commands.bot_has_permissions(manage_messages=True)
+	@commands.has_permissions(manage_messages=True)
+	async def whitelist(self, ctx, user: discord.Member):
+		"""
+		Whitelists a user from the softlock in the current channel.
+		"""
+		db = await readDB()
+		if not str(ctx.channel.id) in db["softlocked_channels"]:
+			return await ctx.send("This channel is not softlocked")
+		elif not db["softlocked_channels"][str(ctx.channel.id)]["user"] == str(ctx.author.id):
+			return await ctx.send("You did not softlock this channel")
+		else:
+			db["softlocked_channels"][str(ctx.channel.id)]["whitelist"].append(user.id)
+			await writeDB(db)
+			return await ctx.send(f"Successfully whitelisted {str(user)}")
+
+	@commands.command(name="unsoftlock", aliases=['unlock', 'usl'])
+	@commands.bot_has_permissions(manage_messages=True)
+	@commands.has_permissions(manage_messages=True)
+	async def unsoftlock(self, ctx, channel: discord.TextChannel=None):
+		"""
+		Unsoftlocks a channel.
+		"""
+		db = await readDB()
+		ch = channel or ctx.channel
+		if str(ch.id) in db["softlocked_channels"]:
+			if db["softlocked_channels"][str(ch.id)]["user"] == str(ctx.author.id):
+				del db["softlocked_channels"][str(ch.id)]
+				await writeDB(db)
+				await ctx.send("Successfully unsoftlocked.")
+			else:
+				await ctx.send("You cannot unlock this, as you are not the one who locked it.")
+		else:
+			await ctx.send("Channel is not softlocked.")
 	
 	@commands.command(name="ban")
 	@commands.bot_has_permissions(ban_members=True)
